Feat: display cards on screen
Utilize useEffect hook to fetch api data. Then use Promise.all to
gather all cards data in an array. Now use a "then" to set the state
variable and trigger a re-render.

Note: trying to assign data outside of a promise may lead to bugs
(not rendering data properly).

The App component returns a list of Card components, each one with a
onClick function attached to the div. This function shuffles the cards
at the moment. Later it will check if the user clicked the same card
twice and doing something based on this check.

Added a few lines to card.css to give an idea about the final page
look.