Feat: let player play the game
The game now has a score and best score. These are added through the
new Scoreboard component.

The Card component now accepts a callback that handles each player
round. This callback is defined in the App.jsx to have access to state
variables.

In the App.jsx, there are new functions to handle each player round, as
said in the previous paragraph. Also there is a new function declared
inside the useEffect hook to get unique ids when fetching API data.

There are a few changes in some .css files to alter the page style and
avoid stretching the pokemon sprite. More changes will be added later.